requirements.txt
boto3
botocore
streamlit
pymupdf
faiss-cpu
numpy
python-dotenv
tqdm
